requirements.txt
pathlib
google-cloud-aiplatform 
PyPDF2 
ratelimit 
backoff 
streamlit
logging
stqdm 
